refactor(core): move local-LLM server shapes to local-llm-servers.ts
Review round 2: local-llm.ts is under a file-size ratchet (ceiling 1481)
and the probe-table addition pushed it to 1501. The server shapes, probe
ordering, and their two helpers now live in local-llm-servers.ts (no
dependency on the client); local-llm.ts re-exports LocalLlmType.
